Remove the utils function.
